requirements.txt
flask
openai
python-dotenv
requests
gTTS
moviepy
Pillow
google-generativeai
